modularize donation feature

+++ client/lib/src/features/purchase/data/repository/purchase_repository.dart
@@ -3,6 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart';
 import 'package:intola/src/features/cart/data/repository/cart_repository.dart';
 import 'package:intola/src/features/cart/domain/model/product_item_model.dart';
+import 'package:intola/src/features/donation/domain/model/donation_model.dart';
 import 'package:intola/src/features/product/domain/model/product_model.dart';
 import 'package:intola/src/features/purchase/domain/model/purchase_history_model.dart';
 import 'package:intola/src/features/purchase/data/network/purchase_network_helper.dart';
@@ -19,6 +20,7 @@ class PurchaseHistoryRepository {
 
   final SecureStorage secureStorage;
   final CartRepository cartRepository;
+  final List<ProductModel> _products = [];
 
   Future<List<PurchaseHistoryModel>> fetchPurchaseHistory() async {
     try {
@@ -49,28 +51,49 @@ class PurchaseHistoryRepository {
     }
   }
 
-  Future<void> addPurchaseHistory() async {
+  Future<void> addPurchaseHistory({DonationModel? donationProduct}) async {
     try {
-      Map<String, ProductItem>? shoppingCart = await cartRepository.fetchCart();
+      if (donationProduct == null) {
+        Map<String, ProductItem>? shoppingCart =
+            await cartRepository.fetchCart();
 
-      if (shoppingCart == null) return;
+        if (shoppingCart == null) return;
 
-      List<ProductModel> cartProducts = [];
+        for (var entry in shoppingCart.entries) {
+          _products.add(entry.value.productModel);
+        }
+        //add purchases to db
+        await purchaseHistoryNetworkHelper.addPurchaseHistory(
+          products: _products,
+        );
 
-      for (var entry in shoppingCart.entries) {
-        cartProducts.add(entry.value.productModel);
+        //clear cart after purchases
+        await secureStorage.delete(key: 'cart');
+      } else {
+        _addDonationProductToPurchaseHistory(donationProduct);
       }
-      //add purchases to db
-      await purchaseHistoryNetworkHelper.addPurchaseHistory(
-        products: cartProducts,
-      );
-
-      //clear cart after purchases
-      await secureStorage.delete(key: 'cart');
     } on Response catch (response) {
       final responseBody = RequestResponse.requestResponse(response);
       return jsonDecode(responseBody);
     }
+  }
+
+  Future<void> _addDonationProductToPurchaseHistory(
+      DonationModel donation) async {
+    _products.add(
+      ProductModel(
+        id: donation.id!,
+        name: donation.name,
+        image: donation.image,
+        price: donation.price,
+        slashprice: donation.price,
+        description: donation.description,
+        quantity: donation.quantity,
+      ),
+    );
+    await purchaseHistoryNetworkHelper.addPurchaseHistory(
+      products: _products,
+    );
   }
 }
